Add generic entrypoint for billing sync
