+++ docs/Setup.md
@@ -300,126 +300,8 @@ from mountaineer_billing import get_billing_router
 controller.app.include_router(get_billing_router())
 ```
 
-## CLI
-
-```python
-from click import command, option
-
-from mountaineer_billing.cli.sync_up import BillingSync
-
-@command()
-@option("--dry-run", is_flag=True, default=False)
-@async_to_sync
-async def sync_plans(dry_run: bool):
-    """
-    Sync local plans to Stripe.
-    
-    This will:
-    1. Fetch all remote products and prices from Stripe
-    2. Calculate differences between local and remote definitions
-    3. Show a detailed summary of proposed changes
-    4. Ask for user confirmation before making any changes (unless --dry-run)
-    5. Apply the changes to Stripe and store price mappings locally
-
-    Use --dry-run to preview what would change without making any changes.
-    """
-    config = AppConfig()  # type: ignore
-
-    async def sync_products(
-        db_connection: DBConnection = Depends(DatabaseDependencies.get_db_connection),
-    ):
-        billing_sync = BillingSync(config=config)
-        sync_diff = await billing_sync.sync_products(
-            products=PRODUCTS, 
-            db_connection=db_connection,
-            dry_run=dry_run
-        )
-        
-        if dry_run:
-            # Show summary when in dry run mode
-            print("\n🔍 DRY RUN SUMMARY:")
-            print(f"  • Would create {len(sync_diff.products_to_create)} products in Stripe")
-            print(f"  • Would update {len(sync_diff.products_to_update)} products in Stripe")
-            print(f"  • Would create {len(sync_diff.prices_to_create)} prices in Stripe")
-            print(f"  • Would store {len(sync_diff.prices_to_store_locally)} price mappings locally")
-            print("\nRun without --dry-run to apply these changes.")
-
-    async with get_function_dependencies(callable=sync_products) as values:
-        await sync_products(**values)
-
-```
-
-The new sync system provides:
-
-- **Diff Calculation**: Compares local and remote definitions to show exactly what will change
-- **User Confirmation**: Interactive prompts before making any changes (batched to minimize interruptions)
-- **Dry Run Mode**: Use `--dry-run` flag to preview changes without making them
-- **Verbose Logging**: Detailed progress and summary information
-- **Safe Operations**: No automatic deletion of existing data
-- **Duplicate Prevention**: Won't create duplicate price mappings in the database
-
-### Sync Workflow
-
-1. **Fetch Remote Data**: Retrieves all existing products and prices from Stripe
-2. **Calculate Differences**: Compares local product definitions with remote state
-3. **Log Summary**: Shows detailed breakdown of what will be created, updated, or stored
-4. **User Confirmation**: Asks for approval before making changes (unless `--dry-run`)
-5. **Apply Changes**: Creates/updates products in Stripe and stores price mappings locally
-
-### Usage Examples
-
-```bash
-# Preview changes without making them
-billing-sync up --config myapp.config:AppConfig --dry-run
-
-# Apply changes with user confirmation
-billing-sync up --config myapp.config:AppConfig
-```
-
-### Example Output
-
-**Dry Run Mode:**
-```
-INFO - Starting billing sync process...
-INFO - Fetching remote products and prices from Stripe...
-INFO - Found 2 remote products and 3 remote prices
-INFO - Processing local product: MYAPP_BRONZE
-INFO - === SYNC SUMMARY ===
-INFO - Products to CREATE remotely (0):
-INFO - Products to UPDATE remotely (0):
-INFO - Prices to CREATE remotely (0):
-INFO - Prices to STORE locally (3):
-INFO -   - MYAPP_BRONZE.DEFAULT: price_def456
-INFO - === END SYNC SUMMARY ===
-INFO - Dry run mode - no changes will be made
-
-🔍 DRY RUN SUMMARY:
-  • Would create 0 products in Stripe
-  • Would update 0 products in Stripe
-  • Would create 0 prices in Stripe
-  • Would store 3 price mappings locally
-
-Run without --dry-run to apply these changes.
-```
-
-**Normal Mode:**
-```
-INFO - Starting billing sync process...
-INFO - Fetching remote products and prices from Stripe...
-INFO - Found 2 remote products and 3 remote prices
-INFO - === SYNC SUMMARY ===
-INFO - Prices to STORE locally (3):
-INFO -   - MYAPP_BRONZE.DEFAULT: price_def456
-INFO -   - MYAPP_SILVER.DEFAULT: price_ghi789
-INFO -   - MYAPP_GOLD.DEFAULT: price_jkl012
-INFO - === END SYNC SUMMARY ===
-INFO - The following changes will be made:
-INFO -   - Store 3 price mappings in local database
-
-Do you want to proceed with these changes? (y/N): y
-INFO - Storing price mapping locally: MYAPP_BRONZE.DEFAULT -> price_def456
-INFO - Billing sync completed successfully
-```
+CLI usage for `billing-sync up` and `billing-sync down` is documented in the
+README.
 
 ## Stripe
 
@@ -439,10 +321,3 @@ But there's no harm in selecting all `customer.subscription` and `checkout.sessi
 ![](./endpoint_add.png)
 
 Once you've created the webhook, you'll need to copy the signing secret and paste it into your env configuration as `STRIPE_WEBHOOK_SECRET`.
-
-In your `pyproject.toml`, you'll need to add the following:
-
-```toml
-[project.scripts]
-billing-sync = "mountaineer_billing.cli:billing_sync"
-```
