--- docs/UI.md
@@ -1,3 +1,0 @@
-# UI
-
-Let's walk through building a flow to allow users to register and manage their subscriptions. We'll use Stripe Checkout to facilitate the purchases, then link back to our webapp to manage.
